+++ apps/editor-dev/pages/plugins/RightPlugin.jsx
@@ -1,19 +1,7 @@
-import { Tab, TabItem } from "@elf-framework/ui";
+import { RightTab } from "./right-panels/RightTab";
 
 export function RightPlugin(editor) {
   editor.registerGroupUI("right-panel", {
-    tabs: (
-      <Tab activeKey={1} full={true}>
-        <TabItem title="Style" key={1}>
-          Style
-        </TabItem>
-        <TabItem title="Inspector" key={2}>
-          Inspector
-        </TabItem>
-        <TabItem title="Code" key={3}>
-          Code
-        </TabItem>
-      </Tab>
-    ),
+    tabs: <RightTab />,
   });
 }