+++ apps/ui-dev/component/color/BaseView.jsx
@@ -1,9 +1,11 @@
+import { useComponentRender } from "@elf-framework/sapa";
 import { Grid } from "@elf-framework/ui";
 
 import { getThemeTokens } from "../../utils/theme";
 import { ColorDetailView } from "./ColorDetailView";
 
 export function BaseView() {
+  useComponentRender("view-mode");
   const tokens = getThemeTokens();
 
   return (