+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 0.0.16
+
+### Patch Changes
+
+- update button variant color
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.16
+
 ## 0.0.13
 
 ### Patch Changes