+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/design-tokens
 
+## 0.0.16
+
+### Patch Changes
+
+- update button variant color
+
 ## 0.0.13
 
 ### Patch Changes